refactor: move application review components to own folder
